# src/Services/ProductService/ProductService.Application/Products/Dtos/ProductDto.cs
namespace ProductService.Application.Products.Dtos;

public record ProductDto
{
    public int Id { get; init; }
    public string Name { get; init; } = string.Empty;
    public string Description { get; init; } = string.Empty;
    public decimal Price { get; init; }
    public int Stock { get; init; }
}